+++ tests/unit/compile/cases/syntax-errors/04.text-php-tag-short.t
@@ -1,3 +1,3 @@
 foo <?= $secret ?>
 ---- THROW% ----
-Unexpected "?"; expected: "!" or <HtmlNameWord> near `?= $secret ?>` in `%s` at line 1
+Unexpected "?"; expected: "!" or <HtmlNameWord> near `?= $secret ?>`