+++ tests/unit/compile/cases/syntax-errors/06.doctype-lt.t
@@ -1,3 +1,3 @@
 <!DOCTYPE foo bar <div> test
 ---- THROW% ----
-Unexpected "<"; expected: ">", "'", '"' or <HtmlNameWord> near `<div> test` in `%s` at line 1
+Unexpected "<"; expected: ">", "'", '"' or <HtmlNameWord> near `<div> test`