+++ tests/unit/compile/cases/syntax-errors/13.html-attribute-duplicate.t
@@ -1,4 +1,4 @@
 Lorem <img src="" Src="" /> ipsum
 <img src="" alt="" src=""/> dolor
 ---- THROW% ----
-HTML attribute `src` is duplicated near `/> dolor` in `%s` at line 2
+HTML attribute `src` is duplicated near `/> dolor`