+++ tests/unit/compile/cases/syntax-errors/20.instruction.unknown.t
@@ -1,3 +1,3 @@
 Lorem { %foobar } ipsum
 ---- THROW% ----
-Unknown instructions `foobar` near `foobar } ipsum` in `%s` at line 1
+Unknown instructions `foobar` near `foobar } ipsum`